Move JPMS integration to a new module
