+++ jpms/src/test/java/module-info.java
@@ -26,30 +26,22 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-import cloud.piranha.http.api.HttpServer;
-import cloud.piranha.webapp.api.WebApplicationExtension;
-
 /**
- * The Piranha Server module.
+ * The Piranha - JPMS integration module.
  *
- * @author Manfred Riem (mriem@manorrock.com)
+ * <p>
+ *  This module delivers the default implementation of the {@link java.lang.module} API.
+ * </p>
+ *
+ * @author Thiago Henrique Hupner
  */
-module cloud.piranha.server {
+module cloud.piranha.jpms {
 
-    exports cloud.piranha.server;
-    requires cloud.piranha.api;
-    requires cloud.piranha.extension.servlet;
-    requires cloud.piranha.http.api;
-    requires cloud.piranha.http.webapp;
-    requires cloud.piranha.jpms;
-    requires cloud.piranha.naming.api;
-    requires cloud.piranha.naming.impl;
-    requires cloud.piranha.naming.thread;
-    requires cloud.piranha.resource;
+    exports cloud.piranha.jpms;
+
     requires cloud.piranha.resource.api;
-    requires cloud.piranha.webapp.api;
-    requires cloud.piranha.webapp.impl;
     requires java.logging;
-    uses HttpServer;
-    uses WebApplicationExtension;
+}
+
+module cloud.piranha.jpms.tests {
 }
