+++ micro/pom.xml
@@ -76,6 +76,12 @@
 
         <!-- -->
 
+        <dependency>
+            <groupId>cloud.piranha</groupId>
+            <artifactId>jpms</artifactId>
+            <version>${project.version}</version>
+            <optional>true</optional>
+        </dependency>
 
         <dependency>
             <groupId>cloud.piranha.resource</groupId>
