+++ server/src/main/java/cloud/piranha/server/ServerPiranha.java
@@ -51,7 +51,7 @@ import cloud.piranha.extension.servlet.ServletExtension;
 import cloud.piranha.http.api.HttpServer;
 import cloud.piranha.http.webapp.HttpWebApplicationServer;
 import cloud.piranha.naming.thread.ThreadInitialContextFactory;
-import cloud.piranha.resource.DefaultModuleFinder;
+import cloud.piranha.jpms.DefaultModuleFinder;
 import cloud.piranha.resource.DirectoryResource;
 import cloud.piranha.webapp.api.WebApplicationExtension;
 import cloud.piranha.webapp.api.WebApplicationServerRequestMapper;
@@ -285,7 +285,7 @@ public class ServerPiranha implements Piranha, Runnable {
     }
 
     private void setupLayers(DefaultWebApplicationClassLoader classLoader) {
-        DefaultModuleFinder defaultModuleFinder = new DefaultModuleFinder(classLoader.getResourceManager().getResourceList());
+        ModuleFinder defaultModuleFinder = new DefaultModuleFinder(classLoader.getResourceManager().getResourceList());
 
         List<String> roots = defaultModuleFinder.findAll()
                 .stream()
